clear search filters component
